remove now-useless configuration variables
The CHOICE environment variable is required when the 'choice' library
is linked with the binary. This was the case for all benchmarks
before, adding a requirement to provide this variable in benchmark
that would not depend on it. The new build system only links 'choice'
when necessary, so we can drop these useless settings.