refactor alloc_slot_slow and find_available_pool
Now find_available_pool is in charge of demoting full pools. This
simplifies the code, and it will make it easier to add support for
delete_remote. Remote deletions are delayed to a buffer, and we want
to perform them in batch after we enter the slow path, but before
demoting any pool (as it may free slots inside).