Amortize the 'malloc' calls by keeping a list of recently-removed cells
Our lit of recently-removed cells is reused for the next creation
operation, and finally freed at the next GC collection. This means
that the number of 'malloc' calls between two GC collections is not
the total number of created roots, but the peak number of created
roots that are still alive.

This makes a difference for bracketed workflows (roots are allocated
and deallocated quickly), in particular the local_roots benchmark.