move full pools out of the way into a second pool ring
On perm_count we could observe that find_available_pool has to scan all
pools when they are all full, and that this can become a time sink
when the number of pools grows large (when it is not neglectible
compared to the pool size; perm_count uses around 2K pools).

The solution in this PR is to move full pools out of the way in
a separate full_pools ring. find_available_pool is in charge of moving
pools there when they are found to be full. They are moved back to the
main ring on deallocation, if they cross an almost-full threshold.